requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
openai>=1.0.0
tiktoken>=0.7.0
gitpython>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
